requirements.txt
bitsandbytes==0.44.1
google_cloud_aiplatform==1.73.0
google-cloud-vision==3.8.1
gradio==5.6.0
opencv-python==4.10.0.84
scipy==1.14.1
torch==2.5.1
torchvision==0.20.1
transformers==4.46.3
